Corrected timing code for some other things. Moved the frame/render rate specifiers to be inlined in the TLTime header so they can be effeiciently used in defines (which were previously just 60.f)
Added a Rotate() function to Type2's to rotate a 2D vector.

Some logical speed ups to quadtree zones which are quite significant. (Reducing array traversal really).

Reduced processing for physics when there is no movement. Changed the transform changes to bits for a very minor speed up.
Removed the rendernode zone box calculation for a speed up.

+++ Code/TootleCore/TLCore.cpp
@@ -133,7 +133,7 @@ SyncBool TLCore::Update()
 	}
 
 	//	turn time since update into 1/60th of a second
-	float Timestep = ((float)TimeSinceUpdate) / (1000.f / (float)TLTime::GetUpdatesPerSecond());
+	float Timestep = ((float)TimeSinceUpdate) / (1000.f / TLTime::GetUpdatesPerSecond());
 
 	//	skip unneccasasry updates
 	if ( Timestep < 1.0f )
